organize completion code in a separate file
